requirements.txt
fastapi==0.115.7 
uvicorn>=0.34.0

Telethon==1.38.1 

python-dotenv==1.0.1
requests==2.32.3

pydantic==2.10.6
PyYAML==6.0.2 

SQLAlchemy==2.0.34 
psycopg2==2.9.9 

langchain==0.3.18
langchain-core==0.3.35 
langchain-google-genai==2.0.9
langgraph==0.2.72

